warning fix updates and a few minor fixes
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@821 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/dbsupport/TableModel2Excel.java
@@ -105,11 +105,11 @@ public class TableModel2Excel
                 oddCellStyle.setFont(cellFont);
                 evenCellStyle.setFont(cellFont);
                 
-                evenCellStyle.setFillPattern((short) HSSFCellStyle.SOLID_FOREGROUND);
-                oddCellStyle.setFillPattern((short) HSSFCellStyle.SOLID_FOREGROUND);
+                evenCellStyle.setFillPattern(HSSFCellStyle.SOLID_FOREGROUND);
+                oddCellStyle.setFillPattern(HSSFCellStyle.SOLID_FOREGROUND);
                 
-                oddCellStyle.setFillForegroundColor((short) HSSFColor.WHITE.index);
-                evenCellStyle.setFillForegroundColor((short) HSSFColor.GREY_25_PERCENT.index);
+                oddCellStyle.setFillForegroundColor(HSSFColor.WHITE.index);
+                evenCellStyle.setFillForegroundColor(HSSFColor.GREY_25_PERCENT.index);
                 
                 oddCellStyle.setAlignment(HSSFCellStyle.ALIGN_CENTER);
                 evenCellStyle.setAlignment(HSSFCellStyle.ALIGN_CENTER);
